feat(suite-native): Propagade handleComponent from BottomSheetFlashList, refactoring
