+++ suite-native/atoms/src/index.ts
@@ -19,6 +19,7 @@ export * from './BottomSheetListItem';
 export * from './OrderedListIcon';
 export * from './Sheet/BottomSheet';
 export * from './Sheet/BottomSheetFlashList';
+export * from './Sheet/BottomSheetGrabber';
 export * from './Sheet/useBottomSheetAnimation';
 export * from './Button/Button';
 export * from './Button/IconButton';
